+++ R/recipy-utility.R
@@ -1,0 +1,16 @@
+#' Returns a recipy attached to bllflow if one is not attached constructs one
+#' based on variables sheet
+#' @export
+GetRecipy <- function(){
+  
+}
+#' Sets the recipy for the bllflow object 
+#' TODO add a way to verify that the recipy works with the bllflow object
+#' @export
+SetRecipy <- function(){
+  
+}
+#' Creates recipy from variable sheet
+ConstructRecipy <- function(){
+  
+}